Minor fix Fix gcc warning
svn path=/trunk/koffice/; revision=177945

+++ lib/kotext/kotextformat.cc
@@ -158,7 +158,7 @@ float KoTextFormat::screenPointSize( const KoZoomHandler* zh ) const
 {
     int pointSizeLU = font().pointSize();
     if ( vAlign() != KoTextFormat::AlignNormal )
-        pointSizeLU = ( ( pointSizeLU *relativeTextSize())/** 2 ) / 3*/ );
+        pointSizeLU = (int)( pointSizeLU *relativeTextSize() );
 
     return zh->layoutUnitToFontSize( pointSizeLU, false /* forPrint */ );
 }
@@ -167,7 +167,7 @@ float KoTextFormat::refPointSize() const
 {
     int pointSizeLU = font().pointSize();
     if ( vAlign() != KoTextFormat::AlignNormal )
-        pointSizeLU = ( ( pointSizeLU * relativeTextSize())/** 2 ) / 3*/ );
+        pointSizeLU = (int)( pointSizeLU * relativeTextSize());
     return KoTextZoomHandler::layoutUnitPtToPt( pointSizeLU );
 }
